requirements-dev.txt
-r requirements.txt
django-debug-toolbar
pytest-django
factory-boy